refactor: 검색 관련 디렉터리 생성 (구조 변경)
