+++ src/main/java/org/swyp/dessertbee/common/scheduler/SearchScheduler.java
@@ -3,7 +3,7 @@ package org.swyp.dessertbee.common.scheduler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import org.swyp.dessertbee.common.service.SearchService;
+import org.swyp.dessertbee.search.service.SearchService;
 
 @Service
 @RequiredArgsConstructor
