+++ src/main/java/org/swyp/dessertbee/search/dto/PopularSearchResponse.java
@@ -1,4 +1,4 @@
-package org.swyp.dessertbee.common.dto;
+package org.swyp.dessertbee.search.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
