+++ src/main/java/org/swyp/dessertbee/search/dto/PopularSearchesList.java
@@ -1,8 +1,7 @@
-package org.swyp.dessertbee.common.dto;
+package org.swyp.dessertbee.search.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
