+++ src/main/java/org/swyp/dessertbee/search/repository/UserSearchHistoryRepository.java
@@ -1,4 +1,4 @@
-package org.swyp.dessertbee.common.repository;
+package org.swyp.dessertbee.search.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import org.swyp.dessertbee.common.entity.UserSearchHistory;
+import org.swyp.dessertbee.search.entity.UserSearchHistory;
 
 import java.util.List;
